Convert Vec2i/f etc structs to templated base structs
Now we have Vec2<T>, Vec3<T> and Rect<T>. Also named the aliases a bit more consistently, where an i suffix means int, f for float and d for double (previously f was double)

+++ src/Game/ThingType.h
@@ -74,7 +74,7 @@ private:
 	ColRGBA colour_     = { 170, 170, 180, 255, 0 };
 	int     radius_     = 20;
 	int     height_     = -1;
-	Vec2f   scale_      = { 1., 1. };
+	Vec2d   scale_      = { 1., 1. };
 	bool    angled_     = true;
 	bool    hanging_    = false;
 	bool    shrink_     = false;
