+++ src/MapEditor/Edit/Input.cpp
@@ -165,7 +165,7 @@ bool Input::mouseMove(int new_x, int new_y)
 
 	// Check if we want to start a selection box
 	if (mouse_drag_ == DragType::Selection
-		&& Vec2f(mouse_pos_.x - mouse_down_pos_.x, mouse_pos_.y - mouse_down_pos_.y).magnitude() > 16)
+		&& Vec2d(mouse_pos_.x - mouse_down_pos_.x, mouse_pos_.y - mouse_down_pos_.y).magnitude() > 16)
 	{
 		mouse_state_ = MouseState::Selection;
 		mouse_drag_  = DragType::None;
@@ -173,7 +173,7 @@ bool Input::mouseMove(int new_x, int new_y)
 
 	// Check if we want to start moving
 	if (mouse_drag_ == DragType::Move
-		&& Vec2f(mouse_pos_.x - mouse_down_pos_.x, mouse_pos_.y - mouse_down_pos_.y).magnitude() > 4)
+		&& Vec2d(mouse_pos_.x - mouse_down_pos_.x, mouse_pos_.y - mouse_down_pos_.y).magnitude() > 4)
 	{
 		mouse_state_ = MouseState::Move;
 		mouse_drag_  = DragType::None;
