+++ src/OpenGL/DrawingFTGL.cpp
@@ -253,7 +253,7 @@ FTFont* getFont(Font font)
 // Draws [text] at [x,y]. If [bounds] is not null, the bounding coordinates of
 // the rendered text string are written to it.
 // -----------------------------------------------------------------------------
-void Drawing::drawText(const string& text, int x, int y, ColRGBA colour, Font font, Align alignment, Rectf* bounds)
+void Drawing::drawText(const string& text, int x, int y, ColRGBA colour, Font font, Align alignment, Rectd* bounds)
 {
 	// Get desired font
 	auto ftgl_font = getFont(font);
